requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-CORS==4.0.0
pyotp==2.9.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
email-validator==2.1.0.post1
python-jose==3.3.0
